refactor: App.tsx에서 스타일 관련 코드 Layout 컴포넌트로 빼기
